Ser/requirements.txt
flask==2.2.3
flask-cors==3.0.10
tushare==1.2.89
pandas==1.5.3
python-dotenv==1.0.0
requests==2.28.2
apscheduler==3.10.1 